[refactor] decompose query synthesizer into three-stage pipeline
Split the monolithic query_synthesizer.py into focused modules:
- _types.py: shared data classes (QuerySynthesisResult, RepoSnapshot, etc.)
- _agent.py: AgentRunner wrapper around Claude Agent SDK
- context_loader.py: repo checkout, graph indexing, block sampling
- query_curator.py: query generation with behavioral/non-behavioral modes
- verifier.py: quality validation and blind target-alignment checks

The query_synthesizer.py now serves as a thin orchestration facade.
