+++ codeminer/dataset/synthesize/__init__.py
@@ -9,11 +9,18 @@ from codeminer.dataset.utils import (
     get_prompt_for_query_type,
 )
 
-from .query_synthesizer import ClaudeQuerySynthesizer, QuerySynthesisResult
+from ._types import QuerySynthesisResult
+from .context_loader import ContextLoader
+from .query_curator import QueryCurator
+from .query_synthesizer import ClaudeQuerySynthesizer
+from .verifier import Verifier
 
 __all__ = [
     "ClaudeQuerySynthesizer",
+    "ContextLoader",
+    "QueryCurator",
     "QuerySynthesisResult",
+    "Verifier",
     "CodeLocation",
     "CodeSearchDataset",
     "CodeSearchQuery",
